// MyPlace/src/app/dashboards/model/constants.ts
export enum WidgetCommands {
    save = 1,
    delete = 2,
    add = 3,
    view = 4,
    config = 5,
    show = 6,
    refreshWidgetView = 7,
    refreshDashboardWidgets = 8,
    refreshAfterDrag = 9,
    rowColSpanChanges = 10,
    saveWidgets = 11,
    copyWidget = 12,
    saveProjectViewerState =14
}
export enum WidgetMode {
    show = 1,
    edit = 2,
    add = 3,
    copy = 4
}
export enum DashboardGroup {
    favorite = 0,
    user = 1,
    new = 2
}